♻️ refactor: move useFormControlContext to separate module
