+++ src/index.ts
@@ -6,7 +6,7 @@ export * from './types';
 export * from './useController';
 export * from './useFieldArray';
 export * from './useForm';
-export { FormProvider, useFormContext } from './useFormContext';
+export * from './useFormContext';
 export * from './useFormState';
 export * from './useWatch';
 export * from './utils';
